almost done

+++ app/src/main/java/com/telestapp/rateexchange/core/data/RatesListInfo.kt
@@ -7,5 +7,5 @@ data class RatesListInfo(
 
 data class ExchangeRate(
     val currency: String,
-    val rate: Float
+    val rate: Double
 )
