+++ app/src/main/java/com/telestapp/rateexchange/database/data/RateExchange.kt
@@ -2,5 +2,5 @@ package com.telestapp.rateexchange.database.data
 
 data class RateExchange(
     val currency: String,
-    val exchange: Float
+    val exchange: Double
 )
